Change Synchroscope Format
So originally the lamps are suppose to increase their emissive intensity and then decrease by using coroutines. Increase the brightness works but the problem is when it needs to dim. It started to flicker and dim but that is not what i want. I suspected that "elapsed time" has something to do with it but i tried to solve it but fail.

So i decided to go through another route. It is still in the testing phase but there will be 3 scripts, the On, Off and manager.

The manager will control the two other scripts to turn on and off adjusting the speed of the dim process and restarting the coroutines.